src/requirements.txt
langgraph~=0.2.63
langchain~=0.3.14
langchain_openai~=0.3.0
python-dotenv~=1.0.1
langchain-core~=0.3.30
langchain-openai~=0.3.0
pytest~=8.3.4
fastapi~=0.115.6
pydantic~=2.10.5
uvicorn~=0.34.0